Move mobile telemetry consent into core
